+++ logs/log_2024-04-21_20-39-02.log
@@ -1,0 +1,13 @@
+
+-----
+Test: test_krasnodar_220_volt_api.py::TestTakeCategoryAndProductList::test_take_category_and_product_list (call)
+Time: 2024-04-21 20:39:02.980162
+Request method: POST
+Request URL: https://krasnodar.220-volt.ru/tracker_advcake/
+
+Response code: 200
+Response text: {"pageType":3,"products":[{"id":"735229","name":"Бензопила HYUNDAI Х 5320"},{"id":"735227","name":"Бензопила HYUNDAI Х 3916"},{"name":"Бензопила HYUNDAI Х 4118","id":"735228"},{"name":"Бензопила HYUNDAI X 520","id":"697556"},{"name":"Бензопила HYUNDAI X 420","id":"697555"},{"name":"Бензопила HYUNDAI X 370","id":"697554"}],"user":{"email":"fbdb1d1b18aa6c08324b7d64b71fb76370690e1d"},"currentCategory":{"id":339,"name":"Сад и дача"},"basketProducts":[]}
+Response headers: {'Server': 'nginx', 'Date': 'Sun, 21 Apr 2024 17:39:05 GMT', 'Content-Type': 'application/json; charset=windows-1251', 'Transfer-Encoding': 'chunked', 'Connection': 'keep-alive', 'Vary': 'Accept-Encoding', 'Set-Cookie': 'chkuidsrv=35e2ca3a69fa466aadf3bed420a32884; expires=Sun, 21-Apr-24 18:39:05 GMT; max-age=3600; domain=.220-volt.ru; path=/, site_version=desktop; expires=Thu, 20 Jun 2024 17:39:05 GMT; path=/; domain=.220-volt.ru, site_version=desktop; expires=Thu, 20 Jun 2024 17:39:05 GMT; path=/; domain=.220-volt.ru, telemetryToken=A53WpsneRNDlDxriGk0KwPw12wxc1SRcozBP8Rx23Djo; expires=Fri, 20 Apr 2029 17:39:05 GMT; path=/; domain=.220-volt.ru, backendStart=1713721145403; expires=Fri, 20 Apr 2029 17:39:05 GMT; path=/; domain=.220-volt.ru, backendEnd=1713721145461; expires=Fri, 20 Apr 2029 17:39:05 GMT; path=/; domain=.220-volt.ru, system_id=815ce8d5977cc27c28aa47f3ed243260z1713721145; path=/;Max-Age=1800;Domain=.220-volt.ru', 'Expires': 'Thu, 01 Jan 1970 00:00:00 GMT', 'Access-Control-Allow-Headers': 'Content-Type', 'Pragma': 'no-cache', 'Cache-Control': 'no-store, no-cache, must-revalidate, max-age=0', 'X-Elapsed': '0.0584', 'Access-Control-Allow-Methods': 'OPTIONS, GET, POST, PUT, HEAD, PATCH, DELETE, OPTIONS', 'Access-Control-Allow-Origin': 'https://*220-volt.ru', 'Last-Modified': 'Sun, 21 Apr 2024 17:39:05 GMT', 'RID': 'ZiVPOfbMSrhTfCSHPYMeLwAAABg', 'Content-Encoding': 'gzip'}
+Response cookies: {'chkuidsrv': '35e2ca3a69fa466aadf3bed420a32884', 'site_version': 'desktop', 'telemetryToken': 'A53WpsneRNDlDxriGk0KwPw12wxc1SRcozBP8Rx23Djo', 'backendStart': '1713721145403', 'backendEnd': '1713721145461', 'system_id': '815ce8d5977cc27c28aa47f3ed243260z1713721145'}
+
+-----
