+++ logs/log_2024-04-21_20-39-37.log
@@ -1,0 +1,13 @@
+
+-----
+Test: test_krasnodar_220_volt_api.py::TestTakeCategoryAndProductList::test_take_category_and_product_list (call)
+Time: 2024-04-21 20:39:37.271299
+Request method: POST
+Request URL: https://krasnodar.220-volt.ru/tracker_advcake/
+
+Response code: 200
+Response text: {"products":[{"id":"735229","name":"Бензопила HYUNDAI Х 5320"},{"id":"735227","name":"Бензопила HYUNDAI Х 3916"},{"name":"Бензопила HYUNDAI Х 4118","id":"735228"},{"id":"697556","name":"Бензопила HYUNDAI X 520"},{"name":"Бензопила HYUNDAI X 420","id":"697555"},{"name":"Бензопила HYUNDAI X 370","id":"697554"}],"pageType":3,"basketProducts":[],"user":{"email":"fbdb1d1b18aa6c08324b7d64b71fb76370690e1d"},"currentCategory":{"id":339,"name":"Сад и дача"}}
+Response headers: {'Server': 'nginx', 'Date': 'Sun, 21 Apr 2024 17:39:39 GMT', 'Content-Type': 'application/json; charset=windows-1251', 'Transfer-Encoding': 'chunked', 'Connection': 'keep-alive', 'Vary': 'Accept-Encoding', 'Set-Cookie': 'chkuidsrv=35e2ca3a69fa466aadf3bed420a32884; expires=Sun, 21-Apr-24 18:39:39 GMT; max-age=3600; domain=.220-volt.ru; path=/, site_version=desktop; expires=Thu, 20 Jun 2024 17:39:39 GMT; path=/; domain=.220-volt.ru, site_version=desktop; expires=Thu, 20 Jun 2024 17:39:39 GMT; path=/; domain=.220-volt.ru, telemetryToken=A53WpsneRNDlDxriGk0KwPpkYJIIzGduCH6PGjfyBYlQ; expires=Fri, 20 Apr 2029 17:39:39 GMT; path=/; domain=.220-volt.ru, backendStart=1713721179685; expires=Fri, 20 Apr 2029 17:39:39 GMT; path=/; domain=.220-volt.ru, backendEnd=1713721179750; expires=Fri, 20 Apr 2029 17:39:39 GMT; path=/; domain=.220-volt.ru, system_id=a506273aedf6a1d159d25c34398c9c31z1713721179; path=/;Max-Age=1800;Domain=.220-volt.ru', 'Expires': 'Thu, 01 Jan 1970 00:00:00 GMT', 'Pragma': 'no-cache', 'Access-Control-Allow-Headers': 'Content-Type', 'Access-Control-Allow-Methods': 'OPTIONS, GET, POST, PUT, HEAD, PATCH, DELETE, OPTIONS', 'Access-Control-Allow-Origin': 'https://*220-volt.ru', 'Cache-Control': 'no-store, no-cache, must-revalidate, max-age=0', 'X-Elapsed': '0.0652', 'Last-Modified': 'Sun, 21 Apr 2024 17:39:39 GMT', 'RID': 'ZiVPW-5bN@LXqkueqse@tgAAAAk', 'Content-Encoding': 'gzip'}
+Response cookies: {'chkuidsrv': '35e2ca3a69fa466aadf3bed420a32884', 'site_version': 'desktop', 'telemetryToken': 'A53WpsneRNDlDxriGk0KwPpkYJIIzGduCH6PGjfyBYlQ', 'backendStart': '1713721179685', 'backendEnd': '1713721179750', 'system_id': 'a506273aedf6a1d159d25c34398c9c31z1713721179'}
+
+-----
