+++ logs/log_2024-04-21_20-50-58.log
@@ -1,0 +1,13 @@
+
+-----
+Test: test_krasnodar_220_volt_api.py::TestTakeCategoryAndProductList::test_take_category_and_product_list (call)
+Time: 2024-04-21 20:50:58.859334
+Request method: POST
+Request URL: https://krasnodar.220-volt.ru/tracker_advcake/
+
+Response code: 200
+Response text: {"basketProducts":[],"user":{"email":"fbdb1d1b18aa6c08324b7d64b71fb76370690e1d"},"currentCategory":{"name":"Сад и дача","id":339},"products":[{"id":"735229","name":"Бензопила HYUNDAI Х 5320"},{"id":"735227","name":"Бензопила HYUNDAI Х 3916"},{"name":"Бензопила HYUNDAI Х 4118","id":"735228"},{"id":"697556","name":"Бензопила HYUNDAI X 520"},{"id":"697555","name":"Бензопила HYUNDAI X 420"},{"id":"697554","name":"Бензопила HYUNDAI X 370"}],"pageType":3}
+Response headers: {'Server': 'nginx', 'Date': 'Sun, 21 Apr 2024 17:51:01 GMT', 'Content-Type': 'application/json; charset=windows-1251', 'Transfer-Encoding': 'chunked', 'Connection': 'keep-alive', 'Vary': 'Accept-Encoding', 'Set-Cookie': 'chkuidsrv=35e2ca3a69fa466aadf3bed420a32884; expires=Sun, 21-Apr-24 18:51:01 GMT; max-age=3600; domain=.220-volt.ru; path=/, site_version=desktop; expires=Thu, 20 Jun 2024 17:51:01 GMT; path=/; domain=.220-volt.ru, site_version=desktop; expires=Thu, 20 Jun 2024 17:51:01 GMT; path=/; domain=.220-volt.ru, telemetryToken=A53WpsneRNDlDxriGk0KwPdtszKp4gGe9KobBr5azYkc; expires=Fri, 20 Apr 2029 17:51:01 GMT; path=/; domain=.220-volt.ru, backendStart=1713721861327; expires=Fri, 20 Apr 2029 17:51:01 GMT; path=/; domain=.220-volt.ru, backendEnd=1713721861408; expires=Fri, 20 Apr 2029 17:51:01 GMT; path=/; domain=.220-volt.ru, system_id=8c133672a8d3ec5cc0f83a1d11a6d413z1713721861; path=/;Max-Age=1800;Domain=.220-volt.ru', 'Expires': 'Thu, 01 Jan 1970 00:00:00 GMT', 'Pragma': 'no-cache', 'Access-Control-Allow-Headers': 'Content-Type', 'Access-Control-Allow-Methods': 'OPTIONS, GET, POST, PUT, HEAD, PATCH, DELETE, OPTIONS', 'Access-Control-Allow-Origin': 'https://*220-volt.ru', 'Cache-Control': 'no-store, no-cache, must-revalidate, max-age=0', 'X-Elapsed': '0.0812', 'Last-Modified': 'Sun, 21 Apr 2024 17:51:01 GMT', 'RID': 'ZiVSBWgLC40UYeRVx4BnhQAAACc', 'Content-Encoding': 'gzip'}
+Response cookies: {'chkuidsrv': '35e2ca3a69fa466aadf3bed420a32884', 'site_version': 'desktop', 'telemetryToken': 'A53WpsneRNDlDxriGk0KwPdtszKp4gGe9KobBr5azYkc', 'backendStart': '1713721861327', 'backendEnd': '1713721861408', 'system_id': '8c133672a8d3ec5cc0f83a1d11a6d413z1713721861'}
+
+-----
