+++ tests/test_krasnodar_220_volt_api.py
@@ -1,9 +1,8 @@
+import json
 import allure
+from requests import Response
 from utils.checking import Checking
 from utils.api import Krasnodar220VoltApi
-from utils.pydantic_model import validate_data
-from utils.pydantic_model import Products
-
 
 """Получения информации о категории и продуктах"""
 
@@ -15,12 +14,10 @@ class TestTakeCategoryAndProductList():
         print("Метод POST")
         result_post = Krasnodar220VoltApi.category_and_product_list()
         Checking.check_status_code(result_post, 200)
-
-    def test_validate_data(self):
-        validate_data(Krasnodar220VoltApi.category_and_product_list().json(), Products)
-
-
-
+        # check_post = result_post.json()
+        # Checking.check_json_token(result_post, ['status', 'place_id', 'scope', 'reference', 'id'])
+        # # token = json.loads(result_post.text)
+        # Checking.check_json_value(result_post, 'status', 'OK')
